refactor(backend): price_heal joins background_tasks, market_state joins domain
Two mislabeled 'services':
- price_heal.py is job logic — its only consumer is the scheduler
  (background_tasks/jobs.py), and its module state is scan-cadence
  throttling. Scheduler-only code now lives in app/background_tasks/.
- market_state.py is a pure trait table over Yahoo's six-state
  vocabulary — zero I/O, same nature as domain/instrument.py. Moved to
  app/domain/.

Tests follow: tests/background_tasks/{test_registry,test_price_heal}.py.
No behavior change.
