+++ backend/tests/background_tasks/test_registry.py
